Time: 4 ms (23.18%), Space: 8.4 MB (41.39%) - LeetHub

+++ 0050-powx-n/0050-powx-n.cpp
@@ -1,22 +1,3 @@
-// class Solution {
-// public:
-//     double myPow(double x, int n) {
-//         double ans = 1;
-//         long long nn = n;
-//         if(n<0){
-//            nn = nn*-1;
-//         }
-
-//         for(int i=0;i<nn;i++){
-//             ans = ans*x;
-//         }
-//         if(n<0){
-//             return ((double)(1.0)/ans);
-//         }
-//         return ans;
-//     }
-// };
-
 class Solution {
 public:
     double myPow(double x, int n) {
@@ -36,8 +17,53 @@ public:
             }
         }
         if(n<0){
-            ans = (double)(1.0)/ans;
+            ans = (double)(1.0)/(double)(ans);
         }
         return ans;
     }
 };
+
+
+// class Solution {
+// public:
+//     double myPow(double x, int n) {
+//         double ans = 1.0;
+//         long long nn = n;
+//         if(n<0){
+//            nn = nn*-1;
+//         }
+
+//         for(int i=0;i<nn;i++){
+//             ans = ans*x;
+//         }
+//         if(n<0){
+//             ans = (double)(1.0)/(double)(ans);
+//         }
+//         return ans;
+//     }
+// };
+
+// // class Solution {
+// // public:
+// //     double myPow(double x, int n) {
+// //         long long nn = n;
+// //         double ans = 1.0;
+// //         if(n<0){
+// //             nn = nn*-1;
+// //         }
+// //         while(nn){
+// //             if(nn%2){
+// //                 ans = ans*x;
+// //                 nn = nn-1;
+// //             }
+// //             else{
+// //                 x = x*x;
+// //                 nn = nn/2;
+// //             }
+// //         }
+// //         if(n<0){
+// //             ans = (double)(1.0)/ans;
+// //         }
+// //         return ans;
+// //     }
+// // };
